Move class to another class
